add logging to aggregation
