+++ src/vivarium_inputs/data_artifact/aggregation.py
@@ -7,7 +7,7 @@ from pathlib import Path
 from typing import Set
 
 from vivarium_public_health.dataset_manager import Artifact, ArtifactException, get_location_term
-from vivarium_inputs.data_artifact.utilities import get_versions
+from vivarium_inputs.data_artifact.utilities import get_versions, setup_logging
 
 
 _log = logging.getLogger(__name__)
@@ -22,14 +22,16 @@ def unpack_arguments(parser, rawargs=None):
     parser.add_argument('--config_path', type=str, required=True)
     parser.add_argument('--output_root', type=str, required=True)
     parser.add_argument('--locations', nargs='+')
+    parser.add_argument('--verbose', action='store_true')
     args = parser.parse_args(rawargs)
 
     config_path = Path(args.config_path)
     locations = args.locations
     locations = [l.replace('_', ' ') for l in locations]
     output_root = args.output_root
+    verbose = args.verbose
 
-    return config_path, locations, output_root
+    return config_path, locations, output_root, verbose
 
 
 def aggregate(artifacts: [Artifact], artifact_path: str) -> Artifact:
@@ -56,6 +58,8 @@ def aggregate(artifacts: [Artifact], artifact_path: str) -> Artifact:
     if artifact_path.is_file():
         artifact_path.unlink()
 
+    logging.debug(f'Validating individual artifacts for the common keyspace.')
+
     locations = set().union(*[a.load('metadata.locations') for a in artifacts])
     keyspace_set = set().union(*[a.load('metadata.keyspace') for a in artifacts])
     valid_artifacts, valid_locations = [], []
@@ -64,17 +68,18 @@ def aggregate(artifacts: [Artifact], artifact_path: str) -> Artifact:
             valid_artifacts.append(a)
             valid_locations.extend(a.load('metadata.locations'))
         else:
-            warnings.warn(f'missing_keys: {keyspace_set.difference(set(a.load("metadata.keyspace")))} '
-                          f'for location:{a.load("metadata.locations")} All the artifacts not aggregated'
-                          f'will be stored in {artifact_path}/broken_artifacts.', ArtifactAggregationWarning)
+            logging.warning(f'Missing_keys: {keyspace_set.difference(set(a.load("metadata.keyspace")))} '
+                            f'for location:{a.load("metadata.locations")} All artifacts not aggregated'
+                            f'will be stored in {artifact_path}/broken_artifacts.', ArtifactAggregationWarning)
 
     artifact = Artifact(artifact_path.as_posix())
     artifact.write("metadata.locations", valid_locations)
     current_versions = get_versions()
 
     artifact.write("metadata.versions", current_versions)
-
-    for k in keyspace_set - set(METADATA):
+    keys_to_process = keyspace_set - set(METADATA)
+    for i, k in enumerate(keys_to_process):
+        logging.debug(f'Processing key {i+1} of {len(keys_to_process)}: {k}.')
         data = [a.load(k) for a in valid_artifacts]
         if isinstance(data[0], pd.DataFrame):
             if 'location' in data[0].columns:
@@ -86,6 +91,8 @@ def aggregate(artifacts: [Artifact], artifact_path: str) -> Artifact:
         else:
             assert np.all([d == data[0] for d in data])
             artifact.write(k, data[0])
+
+    logging.debug('Finished aggregating keys. Commencing cleanup of location-specific artifact.')
 
     # clean-up
     for loc in valid_locations:
@@ -103,6 +110,8 @@ def aggregate(artifacts: [Artifact], artifact_path: str) -> Artifact:
             f = artifact_path.parent/f'{artifact_path.stem}_{l}.hdf'
             f.rename(new_dir/f.name)
 
+    logging.debug('Aggregation cleanup finished.')
+
     return artifact
 
 
@@ -113,7 +122,9 @@ def main(rawargs=None):
         in the union of keyspaces.
     """
     parser = argparse.ArgumentParser()
-    config_path, locations, output = unpack_arguments(parser, rawargs)
+    config_path, locations, output, verbose = unpack_arguments(parser, rawargs)
+
+    setup_logging(output, verbose, 'aggregation', config_path, False)
 
     individual_artifacts = []
     for loc in locations:
@@ -122,8 +133,9 @@ def main(rawargs=None):
         individual_artifacts.append(Artifact(individual_path.as_posix()))
 
     artifact_path = f'{output}/{config_path.stem}.hdf'
-    _log.debug(f'{locations}')
+    logging.debug(f'Beginning aggregation for: {locations}.')
     if len(locations) == 1:
+        logging.debug(f'Processing location 1 of 1: {locations[0]}.')
         location = locations[0].replace(' ', '_').replace("'", "-")
         current_artifact = Path(output) / f'{config_path.stem}_{location}.hdf'
         current_artifact.rename(Path(artifact_path))
@@ -133,7 +145,7 @@ def main(rawargs=None):
 
 
 def disaggregate(config_name: str, output_root: str) -> Set:
-    """Disaggreagte the existing multi-location artifacts into the single
+    """Disaggregate the existing multi-location artifacts into the single
     location artifacts for appending. It is only called when the append flag
     is true. If the existing artifact was built under the different versions of
     relevant libraries from the versions of those at the moment when appending
@@ -161,8 +173,8 @@ def disaggregate(config_name: str, output_root: str) -> Set:
 
     if existing_artifact.load('metadata.versions') != get_versions():
         #  FIXME: For now we only warn and build from scratch. We need a smarter way to handle ths.
-        warnings.warn('Your artifact was built under the different versions and we cannot append it. It will be built '
-                      'from scratch', ArtifactAggregationWarning)
+        warnings.warn('Your artifact was built under different versions from those currently installed, '
+                      'so we cannot append to it. It will be built from scratch.', ArtifactAggregationWarning)
 
         initial_artifact_path.unlink()
         existing_locations = {}
